+++ src/junk-filter.lua
@@ -61,6 +61,19 @@ local function getJunkItems(filterFunc, items)
   return items
 end
 
+--- Returns `true` if the given `itemQuality` is enabled within the given `checkBoxValues`.
+--- @param itemQuality integer
+--- @param checkBoxValues ItemQualityCheckBoxValues
+local function isItemQualityCheckBoxValueEnabled(itemQuality, checkBoxValues)
+  return (
+    (checkBoxValues.poor and itemQuality == Enum.ItemQuality.Poor) or
+    (checkBoxValues.common and itemQuality == (Enum.ItemQuality.Common or Enum.ItemQuality.Standard)) or
+    (checkBoxValues.uncommon and itemQuality == (Enum.ItemQuality.Uncommon or Enum.ItemQuality.Good)) or
+    (checkBoxValues.rare and itemQuality == Enum.ItemQuality.Rare) or
+    (checkBoxValues.epic and itemQuality == Enum.ItemQuality.Epic)
+  )
+end
+
 -- ============================================================================
 -- JunkFilter
 -- ============================================================================
@@ -146,27 +159,47 @@ function JunkFilter:IsJunkItem(item)
 
   -- Exclude unbound equipment.
   if currentState.excludeUnboundEquipment and (Items:IsItemEquipment(item) and not Items:IsItemBound(item)) then
-    return false, concat(L.OPTIONS_TEXT, L.EXCLUDE_UNBOUND_EQUIPMENT_TEXT)
+    local checkBoxValues = currentState.itemQualityCheckBoxes.excludeUnboundEquipment
+    if isItemQualityCheckBoxValueEnabled(item.quality, checkBoxValues) then
+      return false, concat(L.OPTIONS_TEXT, L.EXCLUDE_UNBOUND_EQUIPMENT_TEXT)
+    end
   end
 
-  -- Include poor items.
-  if currentState.includePoorItems and item.quality == Enum.ItemQuality.Poor then
-    return true, concat(L.OPTIONS_TEXT, L.INCLUDE_POOR_ITEMS_TEXT)
+  -- Exclude warband equipment.
+  if Addon.IS_RETAIL and currentState.excludeWarbandEquipment and Items:IsItemWarbandEquipment(item) then
+    local checkBoxValues = currentState.itemQualityCheckBoxes.excludeWarbandEquipment
+    if isItemQualityCheckBoxValueEnabled(item.quality, checkBoxValues) then
+      return false, concat(L.OPTIONS_TEXT, L.EXCLUDE_WARBAND_EQUIPMENT_TEXT)
+    end
   end
 
-  -- Soulbound equipment filters.
-  if Items:IsItemBound(item) and Items:IsItemEquipment(item) then
+  -- Include by quality.
+  if currentState.includeByQuality then
+    local checkBoxValues = currentState.itemQualityCheckBoxes.includeByQuality
+    if isItemQualityCheckBoxValueEnabled(item.quality, checkBoxValues) then
+      return true, concat(L.OPTIONS_TEXT, L.INCLUDE_BY_QUALITY_TEXT)
+    end
+  end
+
+  -- Equipment-based include filters.
+  if Items:IsItemEquipment(item) then
     -- Include below item level.
     if currentState.includeBelowItemLevel.enabled then
       local value = currentState.includeBelowItemLevel.value
       if item.itemLevel < value then
-        local valueText = Colors.Grey("(%s)"):format(Colors.Yellow(value))
-        return true, concat(L.OPTIONS_TEXT, L.INCLUDE_BELOW_ITEM_LEVEL_TEXT .. " " .. valueText)
+        local checkBoxValues = currentState.itemQualityCheckBoxes.includeBelowItemLevel
+        if isItemQualityCheckBoxValueEnabled(item.quality, checkBoxValues) then
+          local valueText = Colors.Grey("(%s)"):format(Colors.Yellow(value))
+          return true, concat(L.OPTIONS_TEXT, L.INCLUDE_BELOW_ITEM_LEVEL_TEXT .. " " .. valueText)
+        end
       end
     end
     -- Include unsuitable equipment.
     if currentState.includeUnsuitableEquipment and not Items:IsItemSuitable(item) then
-      return true, concat(L.OPTIONS_TEXT, L.INCLUDE_UNSUITABLE_EQUIPMENT_TEXT)
+      local checkBoxValues = currentState.itemQualityCheckBoxes.includeUnsuitableEquipment
+      if isItemQualityCheckBoxValueEnabled(item.quality, checkBoxValues) then
+        return true, concat(L.OPTIONS_TEXT, L.INCLUDE_UNSUITABLE_EQUIPMENT_TEXT)
+      end
     end
   end