+++ src/ui/main-window/main-window.lua
@@ -1,10 +1,12 @@
 local ADDON_NAME = ... ---@type string
 local Addon = select(2, ...) ---@type Addon
+local Actions = Addon:GetModule("Actions")
 local Colors = Addon:GetModule("Colors")
 local Commands = Addon:GetModule("Commands")
 local L = Addon:GetModule("Locale")
 local Lists = Addon:GetModule("Lists")
 local MainWindowOptions = Addon:GetModule("MainWindowOptions")
+local StateManager = Addon:GetModule("StateManager")
 local Widgets = Addon:GetModule("Widgets")
 
 --- @class MainWindow
@@ -36,7 +38,7 @@ end
 
 MainWindow.frame = (function()
   local NUM_LIST_FRAME_BUTTONS = 7
-  local OPTIONS_FRAME_WIDTH = 250
+  local OPTIONS_FRAME_WIDTH = 275
   local LIST_FRAME_WIDTH = 250
   local TOTAL_FRAME_WIDTH = (
     Widgets:Padding() +
@@ -54,7 +56,14 @@ MainWindow.frame = (function()
     width = TOTAL_FRAME_WIDTH,
     height = 600,
     titleText = Colors.Blue(ADDON_NAME),
+    enableClickHandling = true
   })
+
+  frame:SetClickHandler("RightButton", "SHIFT", function()
+    StateManager:Dispatch(Actions:ResetMainWindowPoint())
+  end)
+
+  Widgets:ConfigureForPointSync(frame, "MainWindow")
 
   -- Version text.
   frame.versionText = frame.titleButton:CreateFontString("$parent_VersionText", "ARTWORK", "GameFontNormalSmall")