+++ src/ui/widgets/_init.lua
@@ -1,4 +1,9 @@
+local ADDON_NAME = ... ---@type string
 local Addon = select(2, ...) ---@type Addon
+local Actions = Addon:GetModule("Actions")
+local E = Addon:GetModule("Events")
+local EventManager = Addon:GetModule("EventManager")
+local StateManager = Addon:GetModule("StateManager")
 
 --- @class Widgets
 local Widgets = Addon:GetModule("Widgets")
@@ -27,4 +32,57 @@ do -- Widget:Padding()
     end
     return value
   end
+end
+
+do -- Widget:GetUniqueName()
+  local ids = {}
+
+  --- Returns a unique variant of the given `widgetName`.
+  --- @param widgetName string
+  --- @return string
+  function Widgets:GetUniqueName(widgetName)
+    ids[widgetName] = (ids[widgetName] or 0) + 1
+    return ("%s_%s%s"):format(ADDON_NAME, widgetName, ids[widgetName])
+  end
+end
+
+--- Configures a draggable `frame` to refresh or save its point based on certain events.
+--- @param frame Frame | any
+--- @param stateType "MainWindow" | "JunkFrame" | "TransportFrame" | "MerchantButton"
+function Widgets:ConfigureForPointSync(frame, stateType)
+  local getPoint, setPoint
+
+  if stateType == "MainWindow" then
+    getPoint = function() return StateManager:GetGlobalState().points.mainWindow end
+    setPoint = function(point) StateManager:Dispatch(Actions:SetMainWindowPoint(point)) end
+  elseif stateType == "JunkFrame" then
+    getPoint = function() return StateManager:GetGlobalState().points.junkFrame end
+    setPoint = function(point) StateManager:Dispatch(Actions:SetJunkFramePoint(point)) end
+  elseif stateType == "TransportFrame" then
+    getPoint = function() return StateManager:GetGlobalState().points.transportFrame end
+    setPoint = function(point) StateManager:Dispatch(Actions:SetTransportFramePoint(point)) end
+  elseif stateType == "MerchantButton" then
+    getPoint = function() return StateManager:GetGlobalState().points.merchantButton end
+    setPoint = function(point) StateManager:Dispatch(Actions:SetMerchantButtonPoint(point)) end
+  end
+
+  local function refresh()
+    local p = getPoint()
+    frame:ClearAllPoints()
+    frame:SetPoint(p.point, nil, p.relativePoint, p.offsetX, p.offsetY)
+  end
+
+  local function save()
+    local point, _, relativePoint, offsetX, offsetY = frame:GetPoint()
+    setPoint({
+      point = point,
+      relativePoint = relativePoint,
+      offsetX = offsetX,
+      offsetY = offsetY
+    })
+  end
+
+  EventManager:On(E.StateUpdated, refresh)
+  frame:HookScript("OnShow", refresh)
+  frame:HookScript("OnDragStop", save)
 end