+++ src/ui/widgets/button.lua
@@ -22,47 +22,64 @@ local Widgets = Addon:GetModule("Widgets")
 --- @return ButtonWidget frame
 function Widgets:Button(options)
   -- Defaults.
+  options.name = Addon:IfNil(options.name, Widgets:GetUniqueName("Button"))
   options.frameType = "Button"
   options.labelColor = Addon:IfNil(options.labelColor, Colors.Gold)
 
   ---@class ButtonWidget : FrameWidget, Button
   local frame = self:Frame(options)
   frame.onClick = options.onClick
-  frame:SetBackdropColor(Colors.DarkGrey:GetRGBA(0.75))
-  frame:SetBackdropBorderColor(0, 0, 0, 1)
+
+  -- Background texture.
+  frame.background = frame:CreateTexture("$parent_Background", "BACKGROUND", nil, -8)
+  frame.background:SetColorTexture(Colors.Backdrop:GetRGBA(1))
+  frame.background:SetAllPoints()
 
   -- Label text.
   frame.label = frame:CreateFontString("$parent_Label", "ARTWORK", "GameFontNormal")
   frame.label:SetText(options.labelText)
-  frame.label:SetTextColor(options.labelColor:GetRGB())
   frame.label:SetPoint("LEFT", frame, self:Padding(0.5), 0)
   frame.label:SetPoint("RIGHT", frame, -self:Padding(0.5), 0)
   frame.label:SetWordWrap(false)
   frame:SetFontString(frame.label)
+  frame:SetHeight(frame.label:GetHeight() + Widgets:Padding(2))
 
-  -- OnClick.
+  local function setNormalColors()
+    frame:SetBackdropColor(Colors.DarkGrey:GetRGBA(0.75))
+    frame:SetBackdropBorderColor(Colors.Black:GetRGBA(1))
+    frame.label:SetTextColor(options.labelColor:GetRGBA(1))
+  end
+
+  local function setHighlightColors()
+    frame:SetBackdropColor(options.labelColor:GetRGBA(0.25))
+    frame:SetBackdropBorderColor(options.labelColor:GetRGBA(1))
+    frame.label:SetTextColor(Colors.White:GetRGBA(1))
+  end
+
+  local function setDisabledColors()
+    frame:SetBackdropColor(Colors.DarkGrey:GetRGBA(0.5))
+    frame:SetBackdropBorderColor(Colors.Black:GetRGBA(1))
+    frame.label:SetTextColor(Colors.Grey:GetRGBA(0.75))
+  end
+
+  -- Initialize colors.
+  setNormalColors()
+
+  -- Scripts.
   frame:SetScript("OnClick", function(self, button)
     if self.onClick then self.onClick(self, button) end
   end)
 
-  -- OnEnter.
-  frame:HookScript("OnEnter", function(self)
-    self:SetBackdropColor(options.labelColor:GetRGBA(0.25))
-    self:SetBackdropBorderColor(options.labelColor:GetRGB())
-    self.label:SetTextColor(Colors.White:GetRGB())
-  end)
+  frame:HookScript("OnEnter", setHighlightColors)
+  frame:HookScript("OnLeave", setNormalColors)
 
-  -- OnLeave.
-  frame:HookScript("OnLeave", function(self)
-    self:SetBackdropColor(Colors.DarkGrey:GetRGBA(0.75))
-    self:SetBackdropBorderColor(0, 0, 0, 1)
-    self.label:SetTextColor(options.labelColor:GetRGB())
-  end)
-
-  -- OnUpdate.
-  frame:SetScript("OnUpdate", function(self)
-    self:SetHeight(self.label:GetHeight() + Widgets:Padding(2))
-    self:SetAlpha(self:IsEnabled() and 1 or 0.5)
+  frame:HookScript("OnDisable", setDisabledColors)
+  frame:HookScript("OnEnable", function()
+    if frame:IsMouseOver() then
+      setHighlightColors()
+    else
+      setNormalColors()
+    end
   end)
 
   return frame