+++ src/ui/widgets/list-frame.lua
@@ -24,6 +24,7 @@ local Widgets = Addon:GetModule("Widgets")
 --- @return ListFrameWidget frame
 function Widgets:ListFrame(options)
   -- Defaults.
+  options.name = Addon:IfNil(options.name, Widgets:GetUniqueName("ListFrame"))
   options.titleText = options.list.name
   options.titleJustify = "LEFT"
 
@@ -37,6 +38,7 @@ function Widgets:ListFrame(options)
   end
 
   function options.itemButtonOnUpdateTooltip(self, tooltip)
+    tooltip:SetOwner(self, "ANCHOR_RIGHT")
     tooltip:SetHyperlink(self.item.link)
     tooltip:AddLine(" ")
     tooltip:AddDoubleLine(L.RIGHT_CLICK, L.REMOVE)