+++ src/ui/widgets/window.lua
@@ -1,6 +1,5 @@
 local Addon = select(2, ...) ---@type Addon
 local Colors = Addon:GetModule("Colors")
-local TickerManager = Addon:GetModule("TickerManager")
 
 --- @class Widgets
 local Widgets = Addon:GetModule("Widgets")
@@ -12,21 +11,6 @@ local Widgets = Addon:GetModule("Widgets")
 --- @class WindowWidgetOptions : TitleFrameWidgetOptions
 
 -- ============================================================================
--- Local Functions
--- ============================================================================
-
-local setFrameLevel
-do
-  local prevLevel = 0
-  setFrameLevel = function(frame)
-    local level = prevLevel + 1
-    prevLevel = level
-    -- Delay to avoid overwrites from existing values in `{character}/layout-local.txt`.
-    TickerManager:After(1, function() frame:SetFrameLevel(level) end)
-  end
-end
-
--- ============================================================================
 -- Window
 -- ============================================================================
 
@@ -35,10 +19,11 @@ end
 --- @return WindowWidget frame
 function Widgets:Window(options)
   -- Defaults.
+  options.name = Addon:IfNil(options.name, Widgets:GetUniqueName("Window"))
   options.points = Addon:IfNil(options.points, { { "CENTER" } })
   options.width = Addon:IfNil(options.width, 675)
   options.height = Addon:IfNil(options.height, 500)
-  options.onUpdateTooltip = nil
+  options.enableDragging = true
   options.titleTemplate = "GameFontNormalLarge"
   options.titleJustify = "LEFT"
 
@@ -46,19 +31,9 @@ function Widgets:Window(options)
   local frame = self:TitleFrame(options)
   frame.titleButton:SetBackdrop(nil)
   frame.titleButton:EnableMouse(false)
-  setFrameLevel(frame)
 
   -- Add as special frame to be hidden on certain events.
   table.insert(UISpecialFrames, frame:GetName())
-
-  -- Make frame moveable.
-  frame:SetFrameStrata("HIGH")
-  frame:SetMovable(true)
-  frame:EnableMouse(true)
-  frame:SetClampedToScreen(true)
-  frame:RegisterForDrag("LeftButton")
-  frame:SetScript("OnDragStart", frame.StartMoving)
-  frame:SetScript("OnDragStop", frame.StopMovingOrSizing)
 
   -- Close button.
   frame.closeButton = self:TitleFrameIconButton({